+++ filesfolders/tests/test_views_api.py
@@ -9,14 +9,14 @@ from test_plus.test import APITestCase
 # Projectroles dependency
 from projectroles.models import SODAR_CONSTANTS
 from projectroles.tests.test_views_api import SODARAPIViewTestMixin
-from projectroles.views_api import (
-    CORE_API_MEDIA_TYPE,
-    CORE_API_DEFAULT_VERSION,
-    INVALID_PROJECT_TYPE_MSG,
-)
+from projectroles.views_api import INVALID_PROJECT_TYPE_MSG
 
 from filesfolders.tests.test_views import ZIP_PATH_NO_FILES, TestViewsBaseMixin
 from filesfolders.models import Folder, File, HyperLink
+from filesfolders.views_api import (
+    FILESFOLDERS_API_MEDIA_TYPE,
+    FILESFOLDERS_API_DEFAULT_VERSION,
+)
 
 
 # SODAR constants
@@ -31,8 +31,8 @@ class TestFilesfoldersAPIViewsBase(
 ):
     """Base class for filesfolders API tests"""
 
-    media_type = CORE_API_MEDIA_TYPE
-    api_version = CORE_API_DEFAULT_VERSION
+    media_type = FILESFOLDERS_API_MEDIA_TYPE
+    api_version = FILESFOLDERS_API_DEFAULT_VERSION
 
     def setUp(self):
         super().setUp()